initramfs-framework: expanding init script
For debugging initramfs start up a helpful utility is the break command
line paremeter that allows the interrupt of the init process after a
module has been ran.

This commit extends the init script to introduce the processing of the
break command line option that will spawn a shell after a module
regardless if the module fails or succeeds.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>